feat : Game 입력값 저장 클래스 생성
